chore: setup alias paths & lefthook
