+++ package.json
@@ -51,6 +51,7 @@
     "@biomejs/biome": "1.8.3",
     "@changesets/cli": "2.27.7",
     "@types/node": "22.4.1",
+    "lefthook": "1.7.14",
     "typescript": "5.5.4",
     "vite": "5.4.1",
     "vite-plugin-dts": "4.0.3",
