+++ src/index.ts
@@ -1,2 +1,3 @@
-export { generateSlug, addLanguagePack } from './generate-slug.js'
+export { generateSlug } from './generate-slug.js'
+export { addLanguagePack } from './add-language-pack.js'
 export * from './types.js'
