+++ src/language-packs/chinese/pack.ts
@@ -1,14 +1,13 @@
 import type { LanguagePack } from '@/types.js'
 import { customRules } from '../custom-rules.js'
 
-export const danishLanguagePack: LanguagePack = {
+export const chineseLanguagePack: LanguagePack = {
   transliterations: {
-    æ: 'ae',
-    Æ: 'Ae',
-    ø: 'oe',
-    Ø: 'Oe',
-    å: 'aa',
-    Å: 'Aa'
+    你: 'ni',
+    好: 'hao'
+  },
+  symbols: {
+    '～': '~'
   },
   customRules: [...customRules]
 }
